Major update with fixed physics
This update fixes the calculation at the Stefan boundary. Previously, it was using a conductive approach, which underestimated the fluxes.
Now, we use the turbulent fluxes directly, along with the conductive fluxes from the firn below. There is also a significant fix for the
surface energy balance for lakes. Previously, surface_fluxes was assuming all of the incoming SW radiation was absorbed even if a lake was present.
In reality, the absorption coefficient for ~60% of the incoming radiation is very small, so only the SWIR and NIR portions are absorbed by the surface
layer(s). Excess radiation is absorbed by the lake, and the remainder (accounting for saturated firn albedo) is included in the Stefan boundary flux calculation as an additional term.

There are various other fixes in here that I will detail in the PR.